feat: 장소 리스트용 지도 컴포넌트를 places 라우트 그룹 layout에 위치
- 장소 리스트 간 원활한 지도 렌더링 제어를 위해 상위 레이아웃에 위치시키기로 결정
- 기존에 embed를 사용하던 search, 장소 상세페이지는 hidden 처리로 레이아웃에 위치한 지도 컴포넌트에 영향을 안 받도록 처리
