+++ app/(lobby)/places/layout.tsx
@@ -1,4 +1,5 @@
-import PlaceListEmbedMap from '@/components/features/place-list/map/PlaceListEmbedMap';
+import PlaceListMap from '@/components/features/place-list/map/PlaceListMap';
+import Script from 'next/script';
 
 export default function PlaceLayout({
   children,
@@ -7,8 +8,12 @@ export default function PlaceLayout({
 }>) {
   return (
     <div className='h-screen w-full flex'>
+      <Script
+        src={`https://maps.googleapis.com/maps/api/js?key=${process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY}&loading=async`}
+        strategy='afterInteractive'
+      />
       {children}
-      <PlaceListEmbedMap />
+      <PlaceListMap />
     </div>
   );
 }
